Modified wrapper functions, added support for sscanf. Removed patch files as per the last commit information.

+++ CC2500_Wireless_UART_Link/STM8S_StdPeriph_Lib/Project/STM8S_StdPeriph_Template/uart.c
@@ -3,7 +3,7 @@
 
 unsigned char index = 0;
 
-void uart_init(int baudrate)
+void uart_init(long baudrate)
 {
   UART1_Cmd(DISABLE);
   UART1_DeInit();
