+++ CC2500_Wireless_UART_Link/STM8S_StdPeriph_Lib/Project/STM8S_StdPeriph_Template/uart.h
@@ -2,7 +2,7 @@
 
 #define UART_H
 
-void uart_init(int);
+void uart_init(long);
 void send_data_uart(char *);
 
 #endif
